[mol|down] Add regression tasks

+++ molecule/preprocess_MoleOOD.py
@@ -23,26 +23,9 @@ if __name__ == "__main__":
         "--datasets",
         nargs="+",
         default=[
-            "hERG",
-            "hERG_Karim",
-            "AMES",
-            "DILI",
-            "Carcinogens_Lagunin",
-            "Tox21",
-            "ClinTox",
-            "PAMPA_NCATS",
-            "HIA_Hou",
-            "Pgp_Broccatelli",
-            "Bioavailability_Ma",
-            "BBB_Martins",
-            "CYP2C19_Veith",
-            "CYP2D6_Veith",
-            "CYP3A4_Veith",
-            "CYP1A2_Veith",
-            "CYP2C9_Veith",
-            "CYP2C9_Substrate_CarbonMangels",
-            "CYP2D6_Substrate_CarbonMangels",
-            "CYP3A4_Substrate_CarbonMangels",
+            "Lipophilicity_AstraZeneca",
+            "Solubility_AqSolDB",
+            "PPBR_AZ",
         ],
         help="the datasets to preprocess",
     )
@@ -61,21 +44,25 @@ if __name__ == "__main__":
     args = parser.parse_args()
     print(args)
     for dataset in args.datasets:
-        result_dir = get_result_dir(dataset)
-        if not os.path.exists(result_dir):
-            os.mkdir(result_dir)
+        if os.path.exists(f"data/{dataset}/moleood/substructures.pkl"):
+            print(f"Dataset {dataset} already preprocessed, skipping")
+            continue
+        if os.path.exists(f"data/{dataset}/smiles.json"):
+            result_dir = get_result_dir(dataset)
+            if not os.path.exists(result_dir):
+                os.mkdir(result_dir)
 
-        with open(f"data/{dataset}/smiles.json", "r") as f:
-            smiles = json.load(f)
+            with open(f"data/{dataset}/smiles.json", "r") as f:
+                smiles = json.load(f)
 
-        file_name = (
-            "substructures.pkl" if args.method == "brics" else "substructures_recap.pkl"
-        )
-        file_name = os.path.join(result_dir, file_name)
-        substruct_list = []
-        for idx, smile in enumerate(tqdm(smiles)):
-            tx = get_substructure(smile=smile, decomp=args.method)
-            substruct_list.append(tx)
+            file_name = (
+                "substructures.pkl" if args.method == "brics" else "substructures_recap.pkl"
+            )
+            file_name = os.path.join(result_dir, file_name)
+            substruct_list = []
+            for idx, smile in enumerate(tqdm(smiles)):
+                tx = get_substructure(smile=smile, decomp=args.method)
+                substruct_list.append(tx)
 
-        with open(file_name, "wb") as Fout:
-            pickle.dump(substruct_list, Fout)
+            with open(file_name, "wb") as Fout:
+                pickle.dump(substruct_list, Fout)
